refactor(desktop): clone PresetsSection into V2PresetsSection
Replaces the v1/v2 adapter hook with a standalone V2PresetsSection
that mirrors PresetsSection file-for-file but reads/writes to the
v2TerminalPresets collection directly. Dumb sub-components
(PresetsTable, PresetEditorSheet, QuickAddPresets) stay imported
from the v1 directory during transition.

This makes v1 deprecation a clean directory delete later instead
of untangling a variant adapter, and lets v2 evolve independently
without worrying about v1 compat.
