+++ apps/desktop/src/renderer/routes/_authenticated/settings/terminal/components/TerminalSettings/TerminalSettings.tsx
@@ -9,6 +9,7 @@ import {
 import { LinkBehaviorSetting } from "./components/LinkBehaviorSetting";
 import { PresetsSection } from "./components/PresetsSection";
 import { SessionsSection } from "./components/SessionsSection";
+import { V2PresetsSection } from "./components/V2PresetsSection";
 
 interface TerminalSettingsProps {
 	visibleItems?: SettingItemId[] | null;
@@ -75,18 +76,28 @@ export function TerminalSettings({
 			</div>
 
 			<SectionList>
-				{(showPresets || showQuickAdd) && (
-					<PresetsSection
-						key="presets"
-						showPresets={showPresets}
-						showQuickAdd={showQuickAdd}
-						variant={isV2CloudEnabled ? "v2" : "v1"}
-						editingPresetId={editingPresetId}
-						onEditingPresetIdChange={onEditingPresetIdChange}
-						pendingCreateProjectId={pendingCreateProjectId}
-						onPendingCreateProjectIdChange={onPendingCreateProjectIdChange}
-					/>
-				)}
+				{(showPresets || showQuickAdd) &&
+					(isV2CloudEnabled ? (
+						<V2PresetsSection
+							key="presets"
+							showPresets={showPresets}
+							showQuickAdd={showQuickAdd}
+							editingPresetId={editingPresetId}
+							onEditingPresetIdChange={onEditingPresetIdChange}
+							pendingCreateProjectId={pendingCreateProjectId}
+							onPendingCreateProjectIdChange={onPendingCreateProjectIdChange}
+						/>
+					) : (
+						<PresetsSection
+							key="presets"
+							showPresets={showPresets}
+							showQuickAdd={showQuickAdd}
+							editingPresetId={editingPresetId}
+							onEditingPresetIdChange={onEditingPresetIdChange}
+							pendingCreateProjectId={pendingCreateProjectId}
+							onPendingCreateProjectIdChange={onPendingCreateProjectIdChange}
+						/>
+					))}
 				{showLinkBehavior && <LinkBehaviorSetting key="link-behavior" />}
 				{showSessions && <SessionsSection key="sessions" />}
 			</SectionList>
