--- apps/desktop/src/renderer/routes/_authenticated/settings/terminal/components/TerminalSettings/components/PresetsSection/hooks/useTerminalPresetsSource/index.ts
@@ -1,5 +1,0 @@
-export {
-	type TerminalPresetsSource,
-	type TerminalPresetsVariant,
-	useTerminalPresetsSource,
-} from "./useTerminalPresetsSource";
