+++ apps/desktop/src/renderer/routes/_authenticated/settings/terminal/components/TerminalSettings/components/V2PresetsSection/index.ts
@@ -1,0 +1,1 @@
+export { V2PresetsSection } from "./V2PresetsSection";
